backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database ORM

SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1

# Environment Variables
python-dotenv==1.0.0

# Authentication & Security
PyJWT==2.11.0
bcrypt==4.1.1
Flask-Login==0.6.3

# IA & Machine Learning
openai==1.11.0
numpy==1.24.3
pandas==2.1.3

# API Validation
Marshmallow==3.20.1
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development Tools
black==23.12.1
flake8==6.1.0
pylint==3.0.3
isort==5.13.2

# Logging & Monitoring
python-logging-loki==0.3.2

# Utility
requests==2.31.0
#Pillow==10.1.0

# WSGI Server
gunicorn==21.2.0
